+++ CHANGELOG.md
@@ -7,12 +7,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [0.3.20] 2026-TODO
 
 - Upgrade YarnSpinner base libraries to 3.2.0. For details, see https://github.com/YarnSpinnerTool/YarnSpinner/releases/tag/v3.2.0
-- Update the source generator to output `.ysls.json` files, used to integrate with other YarnSpinner tools like the VS Code extension.
+- Update the source generator to output `.ysls.json` files, used to integrate with other YarnSpinner tools like the VS Code extension. Updated diagnostics around command registration are also provided.
 - Work around Godot issue where exported C# variables can reset to their initial value: remove initial value from DialogueRunner.dialoguePresenters
 - Register `CustomMarker` and `BasicMarker` as custom types for better editor integration (#119)
 - Fix a bug where the`DialogueRunner.onDialogueComplete` signal was emitted twice when dialogue ends. (fix #120)
 - Add dialogueRunner.allowOptionFallthrough which allows dialogue to proceed when no dialogue options were chosen https://yarnspinner.dev/blog/yarn-spinner-3-1-release/#dialogue-option-fallthrough
-- TODO: more release notes
+- ⚠️ Breaking change: The API for returning diagnostics from ReplacementMarkupHandler implementations has changed. Reference `PaletteMarkerProcessor` for an example of the updates.
+- ⚠️ Breaking change:  Add redesigned IAsyncTypewriter functionality for typewriter animations. BasicTypewriter is removed New types of typewriters are available out of the box: Letter (similar to the prior example BasicTypewriter functionality), Word, and Instant. 
+- Dialogue Presenters now have optional OnNodeEnter and OnNodeExit methods that can be overridden to respond to changes in the running node. 
+- ⚠️ Breaking change: Remove `DialogueViewBase` and `GDScriptPresenterAdapter`, which have been deprecated in favor of DialoguePresenterBase / views implemented directly in GDScript for a while now.
+- LocalizedLine now has a Source property that tells you where the line came from (typically a DialogueRunner instance). https://yarnspinner.dev/blog/yarn-spinner-3-1-release/#lines-know-where-they-came-from
+
 
 ## [0.3.12] 2025-12-21